pdx215: build with lto by default
LTO (Link-Time Optimization) is a compiler optimization technique that allows the compiler to optimize an entire program at link-time, rather than just individual compilation units. When building with LTO, the compiler can analyze the entire program's code and optimize it more effectively, resulting in the fully built ROM having better performance and the binary itself being a smaller size than with it off, only downside being that the build time is longer than with LTO off.
https://llvm.org/docs/LinkTimeOptimization.html
Signed-off-by: saku-bruh <notwolfyy@gmail.com>